django/requirements.txt
Django==5.2.6
djangorestframework==3.16.1
python-dotenv==1.1.1
gunicorn

